requirements.txt
openai
pypdf
tqdm
qdrant-client
langchain
langchain-text-splitters
python-dotenv
pytest